Backend/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
tensorflow
lightgbm
xgboost
joblib
openai
python-dotenv
google-generativeai
